rename test helper
