+++ tests/unit/plugins/modules/test_cpanm.py
@@ -14,7 +14,7 @@ __metaclass__ = type
 
 
 from ansible_collections.community.general.plugins.modules import cpanm
-from .helper import Helper, RunCommandMock
+from .uthelper import UTHelper, RunCommandMock
 
 
-Helper.from_module(cpanm, __name__, mocks=[RunCommandMock])
+UTHelper.from_module(cpanm, __name__, mocks=[RunCommandMock])
