+++ tests/unit/plugins/modules/test_django_command.py
@@ -7,7 +7,7 @@ __metaclass__ = type
 
 
 from ansible_collections.community.general.plugins.modules import django_command
-from .helper import Helper, RunCommandMock
+from .uthelper import UTHelper, RunCommandMock
 
 
-Helper.from_module(django_command, __name__, mocks=[RunCommandMock])
+UTHelper.from_module(django_command, __name__, mocks=[RunCommandMock])
