+++ tests/unit/plugins/modules/test_facter_facts.py
@@ -8,7 +8,7 @@ __metaclass__ = type
 
 
 from ansible_collections.community.general.plugins.modules import facter_facts
-from .helper import Helper, RunCommandMock
+from .uthelper import UTHelper, RunCommandMock
 
 
-Helper.from_module(facter_facts, __name__, mocks=[RunCommandMock])
+UTHelper.from_module(facter_facts, __name__, mocks=[RunCommandMock])
