+++ tests/unit/plugins/modules/test_gconftool2.py
@@ -8,7 +8,7 @@ __metaclass__ = type
 
 
 from ansible_collections.community.general.plugins.modules import gconftool2
-from .helper import Helper, RunCommandMock
+from .uthelper import UTHelper, RunCommandMock
 
 
-Helper.from_module(gconftool2, __name__, mocks=[RunCommandMock])
+UTHelper.from_module(gconftool2, __name__, mocks=[RunCommandMock])
