+++ tests/unit/plugins/modules/test_gio_mime.py
@@ -8,7 +8,7 @@ __metaclass__ = type
 
 
 from ansible_collections.community.general.plugins.modules import gio_mime
-from .helper import Helper, RunCommandMock
+from .uthelper import UTHelper, RunCommandMock
 
 
-Helper.from_module(gio_mime, __name__, mocks=[RunCommandMock])
+UTHelper.from_module(gio_mime, __name__, mocks=[RunCommandMock])
