+++ tests/unit/plugins/modules/test_krb_ticket.py
@@ -8,7 +8,7 @@ __metaclass__ = type
 
 
 from ansible_collections.community.general.plugins.modules import krb_ticket
-from .helper import Helper, RunCommandMock
+from .uthelper import UTHelper, RunCommandMock
 
 
-Helper.from_module(krb_ticket, __name__, mocks=[RunCommandMock])
+UTHelper.from_module(krb_ticket, __name__, mocks=[RunCommandMock])
