+++ tests/unit/plugins/modules/test_xfconf.py
@@ -14,7 +14,7 @@ __metaclass__ = type
 
 
 from ansible_collections.community.general.plugins.modules import xfconf
-from .helper import Helper, RunCommandMock
+from .uthelper import UTHelper, RunCommandMock
 
 
-Helper.from_module(xfconf, __name__, mocks=[RunCommandMock])
+UTHelper.from_module(xfconf, __name__, mocks=[RunCommandMock])
